+++ src/align/bounded/mod.rs
@@ -21,5 +21,8 @@ pub use cloud_search_backward::cloud_search_backward;
 mod cloud_search_forward;
 pub use cloud_search_forward::cloud_search_forward;
 
+mod scoring;
+pub use scoring::{null1_score, null2_score};
+
 mod traceback_bounded;
 pub use traceback_bounded::traceback_bounded;