+++ src/align/bounded/optimal_accuracy_bounded.rs
@@ -7,43 +7,43 @@ pub fn optimal_accuracy_bounded(
     profile: &Profile,
     posterior_matrix: &impl DpMatrix,
     optimal_matrix: &mut impl DpMatrix,
-    params: &RowBounds,
+    row_bounds: &RowBounds,
 ) {
     let end_score: f32 = 1.0;
 
     // initialization of the zero row
-    optimal_matrix.set_special(params.target_start - 1, Profile::SPECIAL_N_IDX, 0.0);
-    optimal_matrix.set_special(params.target_start - 1, Profile::SPECIAL_B_IDX, 0.0);
+    optimal_matrix.set_special(row_bounds.target_start - 1, Profile::SPECIAL_N_IDX, 0.0);
+    optimal_matrix.set_special(row_bounds.target_start - 1, Profile::SPECIAL_B_IDX, 0.0);
     optimal_matrix.set_special(
-        params.target_start - 1,
+        row_bounds.target_start - 1,
         Profile::SPECIAL_E_IDX,
         -f32::INFINITY,
     );
     optimal_matrix.set_special(
-        params.target_start - 1,
+        row_bounds.target_start - 1,
         Profile::SPECIAL_C_IDX,
         -f32::INFINITY,
     );
     optimal_matrix.set_special(
-        params.target_start - 1,
+        row_bounds.target_start - 1,
         Profile::SPECIAL_J_IDX,
         -f32::INFINITY,
     );
 
-    let profile_start_in_first_row = params.left_row_bounds[params.target_start];
-    let profile_end_in_first_row = params.right_row_bounds[params.target_start];
+    let profile_start_in_first_row = row_bounds.left_row_bounds[row_bounds.target_start];
+    let profile_end_in_first_row = row_bounds.right_row_bounds[row_bounds.target_start];
 
     // for profile_idx in 0..=profile.length {
     for profile_idx in (profile_start_in_first_row - 1)..=profile_end_in_first_row {
-        optimal_matrix.set_match(params.target_start - 1, profile_idx, -f32::INFINITY);
-        optimal_matrix.set_insert(params.target_start - 1, profile_idx, -f32::INFINITY);
-        optimal_matrix.set_delete(params.target_start - 1, profile_idx, -f32::INFINITY);
+        optimal_matrix.set_match(row_bounds.target_start - 1, profile_idx, -f32::INFINITY);
+        optimal_matrix.set_insert(row_bounds.target_start - 1, profile_idx, -f32::INFINITY);
+        optimal_matrix.set_delete(row_bounds.target_start - 1, profile_idx, -f32::INFINITY);
     }
 
     // for i in 1..=posterior_matrix.target_length {
-    for target_idx in params.target_start..=params.target_end {
-        let profile_start_in_current_row = params.left_row_bounds[target_idx];
-        let profile_end_in_current_row = params.right_row_bounds[target_idx];
+    for target_idx in row_bounds.target_start..=row_bounds.target_end {
+        let profile_start_in_current_row = row_bounds.left_row_bounds[target_idx];
+        let profile_end_in_current_row = row_bounds.right_row_bounds[target_idx];
 
         optimal_matrix.set_match(target_idx, profile_start_in_current_row - 1, -f32::INFINITY);
         optimal_matrix.set_insert(target_idx, profile_start_in_current_row - 1, -f32::INFINITY);