+++ src/align/bounded/posterior_bounded.rs
@@ -7,33 +7,34 @@ pub fn posterior_bounded(
     forward_matrix: &impl DpMatrix,
     backward_matrix: &impl DpMatrix,
     posterior_matrix: &mut impl DpMatrix,
-    params: &RowBounds,
+    row_bounds: &RowBounds,
 ) {
-    let overall_score: f32 = forward_matrix.get_special(params.target_end, Profile::SPECIAL_C_IDX)
+    let overall_score: f32 = forward_matrix
+        .get_special(row_bounds.target_end, Profile::SPECIAL_C_IDX)
         + profile.special_transition_score(Profile::SPECIAL_C_IDX, Profile::SPECIAL_MOVE_IDX);
 
     let mut denominator: f32;
 
-    posterior_matrix.set_special(params.target_start - 1, Profile::SPECIAL_E_IDX, 0.0);
-    posterior_matrix.set_special(params.target_start - 1, Profile::SPECIAL_N_IDX, 0.0);
-    posterior_matrix.set_special(params.target_start - 1, Profile::SPECIAL_J_IDX, 0.0);
-    posterior_matrix.set_special(params.target_start - 1, Profile::SPECIAL_B_IDX, 0.0);
-    posterior_matrix.set_special(params.target_start - 1, Profile::SPECIAL_C_IDX, 0.0);
+    posterior_matrix.set_special(row_bounds.target_start - 1, Profile::SPECIAL_E_IDX, 0.0);
+    posterior_matrix.set_special(row_bounds.target_start - 1, Profile::SPECIAL_N_IDX, 0.0);
+    posterior_matrix.set_special(row_bounds.target_start - 1, Profile::SPECIAL_J_IDX, 0.0);
+    posterior_matrix.set_special(row_bounds.target_start - 1, Profile::SPECIAL_B_IDX, 0.0);
+    posterior_matrix.set_special(row_bounds.target_start - 1, Profile::SPECIAL_C_IDX, 0.0);
 
-    let profile_start_in_first_row = params.left_row_bounds[params.target_start];
-    let profile_end_in_first_row = params.right_row_bounds[params.target_start];
+    let profile_start_in_first_row = row_bounds.left_row_bounds[row_bounds.target_start];
+    let profile_end_in_first_row = row_bounds.right_row_bounds[row_bounds.target_start];
 
     for profile_idx in (profile_start_in_first_row - 1)..=profile_end_in_first_row {
-        posterior_matrix.set_match(params.target_start - 1, profile_idx, 0.0);
-        posterior_matrix.set_insert(params.target_start - 1, profile_idx, 0.0);
-        posterior_matrix.set_delete(params.target_start - 1, profile_idx, 0.0);
+        posterior_matrix.set_match(row_bounds.target_start - 1, profile_idx, 0.0);
+        posterior_matrix.set_insert(row_bounds.target_start - 1, profile_idx, 0.0);
+        posterior_matrix.set_delete(row_bounds.target_start - 1, profile_idx, 0.0);
     }
 
-    for target_idx in params.target_start..=params.target_end {
+    for target_idx in row_bounds.target_start..=row_bounds.target_end {
         denominator = 0.0;
 
-        let profile_start_in_current_row = params.left_row_bounds[target_idx];
-        let profile_end_in_current_row = params.right_row_bounds[target_idx];
+        let profile_start_in_current_row = row_bounds.left_row_bounds[target_idx];
+        let profile_end_in_current_row = row_bounds.right_row_bounds[target_idx];
 
         posterior_matrix.set_match(target_idx, profile_start_in_current_row - 1, 0.0);
         posterior_matrix.set_insert(target_idx, profile_start_in_current_row - 1, 0.0);